Correção de Bugs
GETS - PROBLEMAS:

~> Get Categorias não faz validações quanto existência ou não de itens (fixed);
~> Get Podcasts não faz validações quanto existência ou não de itens (fixed);
~> Duração do Episódio está com tipo primitivo Double (fixed);
~> Removido Get Episódios - Usuário acessará listagem de episódios no Get ID Podcasts;
~> Data está sendo exibida com hora (VIROU FEATURE);

POST - PROBLEMAS:

~> Post Podcasts recebe internal error ao tentar enviar a requisição com ID de Categoria inexistente (fixed);
~> Post Podcasts retorna Status Ok, porém com categoria e episódios nulos (fixed);

PUTS - PROBLEMAS:

~> Put Categorias quebra ao receber requisição de Id inexistente (fixed);
~> Put Podcasts retornar Status ok, porém com categoria e episódios nulos (fixed);
~> Put Categorias - Data está sendo exibida com hora (VIROU FEATURE);

DELETE - PROBLEMAS:

~> Delete Categorias quebra ao receber requisição de Id inexistente (fixed);
~> Delete Podcasts quebra ao receber requisição de Id inexistente (fixed);
~> Delete Episódios quebra ao receber requisição de Id inexistente (fixed);

+++ GFT-Podcasts/Models/Episodio.cs
@@ -13,7 +13,7 @@ namespace GFT_Podcasts.Models {
 
         public DateTime Lancamento { get; set; }
 
-        public double Duracao { get; set; }
+        public int Duracao { get; set; }
 
         public string LinkAudio { get; set; }
