+++ GFT-Podcasts/Models/ViewModels/EpisodioViewModels/EpisodioSimplificadoViewModel.cs
@@ -10,7 +10,7 @@ namespace GFT_Podcasts.Models.ViewModels.EpisodioViewModels {
 
         public DateTime Lancamento { get; set; }
 
-        public double Duracao { get; set; }
+        public int Duracao { get; set; }
 
         public string LinkAudio { get; set; }
